Backend/requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
black==25.9.0
build==1.3.0
CacheControl==0.14.4
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.3.5
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
deepdiff==8.6.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
email-validator==2.3.0
emoji==2.15.0
faiss-cpu==1.8.0.post1
fastapi==0.119.0
filelock==3.20.0
filetype==1.2.0
firebase_admin==7.1.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.0
google-auth==2.41.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.71.0
gotrue==2.12.4
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.5
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==1.0.0
langchain-core==1.1.0
langchain-google-genai==2.1.12
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.38
loguru==0.7.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orderly-set==5.5.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.3.0
platformdirs==4.5.0
pluggy==1.6.0
postgrest==0.16.11
posthog==5.4.0
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.1.3
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.9
pytokens==0.2.0
PyYAML==6.0.3
RapidFuzz==3.14.1
realtime==1.0.6
referencing==0.37.0
regex==2025.10.23
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
storage3==0.7.7
StrEnum==0.4.15
supabase==2.4.6
supafunc==0.4.7
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.7.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing_extensions==4.15.0
unstructured==0.15.8
unstructured-client==0.25.9
urllib3==2.3.0
uvicorn==0.37.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==12.0
win32_setctime==1.2.0
wrapt==2.0.0
zipp==3.23.0
zstandard==0.25.0
